% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cyclocomp}
\alias{cyclocomp}
\alias{cyclocomp-package}
\title{Cyclomatic Complecity of R Code}
\description{
Cyclomatic complexity is a software metric (measurement), used to indicate
the complexity of a program. It is a quantitative measure of the number of
linearly independent paths through a program's source code. It was developed
by Thomas J. McCabe, Sr. in 1976.
}

