% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dir.R
\name{cyclocomp_package_dir}
\alias{cyclocomp_package_dir}
\title{Cyclomatic complexity of a local package}
\usage{
cyclocomp_package_dir(path = ".", quiet = TRUE)
}
\arguments{
\item{path}{Path to the root directory of the R package.}

\item{quiet}{If `FALSE`, display package build information.}
}
\value{
Data frame with two columns: \code{name} and \code{cyclocomp}.
}
\description{
Automatically builds the package and installs it to a temporary
directory.
}
\seealso{
Other cyclomatic complexity: 
\code{\link{cyclocomp}()},
\code{\link{cyclocomp_package}()}
}
\concept{cyclomatic complexity}
